+++ modules/file_organization_module/file_organization.py
@@ -163,6 +163,7 @@ class FileOrganization(QWidget):
         self.files_segmentation = QCheckBox('Segmentation module output files (*'+gf.output_suffixes['segmentation']+')')
         self.files_cell_tracking = QCheckBox('Cell tracking module output files (*'+gf.output_suffixes['cell_tracking']+')')
         self.files_graph_filtering = QCheckBox('Graph filtering module output files (*'+gf.output_suffixes['graph_filtering']+')')
+        self.files_events_selection = QCheckBox('Events selection module output files (*'+gf.output_suffixes['events_selection']+')')
         self.files_image_cropping = QCheckBox('Image cropping module output files (*'+gf.output_suffixes['image_cropping']+')')
         self.files_ground_truth_generator = QCheckBox('Ground truth generator module output files (*'+gf.output_suffixes['ground_truth_generator']+')')
         self.files_intermediate_logs = QCheckBox('Preserve intermediate log files')
@@ -214,6 +215,7 @@ class FileOrganization(QWidget):
         layout2.addWidget(self.files_segmentation)
         layout2.addWidget(self.files_cell_tracking)
         layout2.addWidget(self.files_graph_filtering)
+        layout2.addWidget(self.files_events_selection)
         layout2.addWidget(self.files_image_cropping)
         layout2.addWidget(self.files_ground_truth_generator)
         layout2.addWidget(self.files_intermediate_logs)
@@ -285,6 +287,8 @@ class FileOrganization(QWidget):
             patterns.append(gf.output_suffixes['cell_tracking']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')
         if self.files_graph_filtering.isChecked():
             patterns.append(gf.output_suffixes['graph_filtering']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')
+        if self.files_events_selection.isChecked():
+            patterns.append(gf.output_suffixes['events_selection']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')
         if self.files_image_cropping.isChecked():
             patterns.append(gf.output_suffixes['image_cropping']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')
 
@@ -352,6 +356,8 @@ class FileOrganization(QWidget):
             patterns.append(gf.output_suffixes['cell_tracking']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')
         if self.files_graph_filtering.isChecked():
             patterns.append(gf.output_suffixes['graph_filtering']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')
+        if self.files_events_selection.isChecked():
+            patterns.append(gf.output_suffixes['events_selection']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')
         if self.files_image_cropping.isChecked():
             patterns.append(gf.output_suffixes['image_cropping']+'[A-Za-z0-9-]*'+'[a-zA-Z4.]*$')