+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add documentation (using MkDocs).
 * Add small sample images and cellpose model fine-tuned for bright-field images of yeast cells.
 * Add simple plugins system.
+* Add a new Events selection module.
 
 ### Changed